Added ability to use internal resetcache endpoint.

+++ PurgeRuleCache/Program.cs
@@ -17,6 +17,7 @@ namespace PurgeRuleCache
             var connectionType = GetConnectionType(args[0]);
             RestClient client;
 
+            var isServiceAccount = true;
             switch (connectionType)
             {
                 case RestClient.ConnectionType.NtlmUserViaProxy:
@@ -30,7 +31,7 @@ namespace PurgeRuleCache
                     client.AsNtlmUserViaProxy();
                     break;
                 case RestClient.ConnectionType.DirectConnection:
-                    if (args.Length != 5)
+                    if (args.Length != 3 && args.Length != 5)
                     {
                         PrintUsage();
                         return;
@@ -44,7 +45,16 @@ namespace PurgeRuleCache
                     }
 
                     var certificates = client.LoadCertificateFromStore();
-                    client.AsDirectConnection(args[3], args[4], port, false, certificates);
+                    var userDir = "INTERNAL";
+                    var userId = "sa_repository";
+                    if (args.Length == 5)
+                    {
+                        userDir = args[3];
+                        userId = args[4];
+                        isServiceAccount = false;
+                    }
+                    Console.WriteLine("Using direct connection as {0}\\{1}", userDir, userId);
+                    client.AsDirectConnection(userDir, userId, port, false, certificates);
                     break;
                 default:
                     PrintUsage();
@@ -56,13 +66,21 @@ namespace PurgeRuleCache
             {
                 client.Get("/qrs/about");
                 Console.WriteLine("Connection successfully established.");
-                Console.Write("Creating dummy rule... ");
-                var response = client.Post("/qrs/systemrule", MakeDummyRuleBody());
-                var id = JObject.Parse(response).Property("id").Value.ToString();
-                Console.WriteLine("Rule successfully created, id=" + id);
-                Console.Write("Deleting rule id=" + id + "... ");
-                client.Delete("/qrs/systemrule/" + id);
-                Console.WriteLine("Rule deleted");
+                if (isServiceAccount)
+                {
+                    Console.WriteLine("Using internal resetcache endpoint.");
+                    client.Post("/qrs/systemrule/security/resetcache", "");
+                }
+                else
+                {
+                    Console.Write("Creating dummy rule... ");
+                    var response = client.Post("/qrs/systemrule", MakeDummyRuleBody());
+                    var id = JObject.Parse(response).Property("id").Value.ToString();
+                    Console.WriteLine("Rule successfully created, id=" + id);
+                    Console.Write("Deleting rule id=" + id + "... ");
+                    client.Delete("/qrs/systemrule/" + id);
+                    Console.WriteLine("Rule deleted");
+                }
             }
             catch (Exception e)
             {
@@ -87,8 +105,9 @@ namespace PurgeRuleCache
         private static void PrintUsage()
         {
             Console.WriteLine("Usage:   PurgeRuleCache.exe --ntlm   <url>");
-            Console.WriteLine("         PurgeRuleCache.exe --direct <url> <port> <userDir> <userId>");
+            Console.WriteLine("         PurgeRuleCache.exe --direct <url> <port> [<userDir> <userId>]");
             Console.WriteLine("Example: PurgeRuleCache.exe --ntlm   https://my.server.url");
+            Console.WriteLine("         PurgeRuleCache.exe --direct https://localhost 4242");
             Console.WriteLine("         PurgeRuleCache.exe --direct https://my.server.url 4242 MyUserDir MyUserId");
             Environment.Exit(1);
         }
